LAN-144 fix: Contato via Forms
