+++ src/app/page.tsx
@@ -15,9 +15,9 @@ export default async function Page({ params }: { params: Params }) {
 	const page = await client
 		.getByUID("landing_page", "landing_page")
 		.catch(()=> notFound());
-	/*const banner = await client
+	const banner = await client
 		.getByUID("banner", "banner")
-		.catch(()=> notFound());*/
+		.catch(()=> notFound());
 	const contato = await client
 		.getByUID("contato", "contato")
 		.catch(()=> notFound());
@@ -27,7 +27,7 @@ export default async function Page({ params }: { params: Params }) {
 		<>
 			{/*Script Hubspot*/}
 			<script type="text/javascript" id="hs-script-loader" async defer src="//js.hs-scripts.com/43688574.js" />
-			{/*<Banner page={banner} />*/}
+			<Banner page={banner} />
 			<Contato page={contato} />
 			<SliceZone slices={page?.data?.slices} components={components} />
 		</>
